refactor: use react-query to update navigation data
